Restored Lara's damage with grenade; HP for statics;

+++ TR5Main/Game/room.h
@@ -44,7 +44,7 @@ typedef struct MESH_INFO
 	short shade;
 	short flags;
 	short staticNumber;
-	int hitPoints;
+	short hitPoints;
 };
 
 typedef struct LIGHTINFO
